+++ downstream-test/src/main/java/react4j/downstream/CollectBuildStats.java
@@ -34,7 +34,7 @@ public final class CollectBuildStats
   {
     Gir.go( () -> {
       final List<String> baseBranches =
-        Arrays.asList( "raw", "arez", "dagger", "raw_maven", "arez_maven", "dagger_maven" );
+        Arrays.asList( "raw", "arez", "dagger", "spritz", "raw_maven", "arez_maven", "dagger_maven" );
 
       final ArrayList<String> branches = new ArrayList<>( baseBranches );
       if ( WorkspaceUtil.buildJ2clBuilds() )